Adding Proxy To React Server
In development mode, the React server runs on port 3000. The
database and backend run on prot 5000. Runing this will cause a
cross origin error in Firefox and Chrome. This "fix" sets the
React server to proxy any requrests that are NOT "text/html" to
port 5000. This works proxy works in development mode. Another
"fix" will have to be applied to production mode. Perhaps using
Express server could offer a solutoin. I'll figure it out.