Created Heatmap To Show Database Connection
I created a HeatMap component that renders in the HomePage. This
new component renders from data fetched from the database. This
works in development mode, because I have a proxy set in package.
json. But, the deatail of a production server proxy still alude me.
Also, the time-series graphs still rely on hardcoded JSON files.